SVG PROTOTYPE with return STRING
