+++ src/main/java/web/SVG/SvgFactory.java
@@ -7,5 +7,5 @@ import domain.Shed.Shed;
  * CREATED BY mathias @ 15-12-2020 - 22:19
  **/
 public interface SvgFactory {
-    void drawCarport(Carport carport, Shed shed);
+    String drawCarport(Carport carport, Shed shed);
 }
