+++ src/test/java/web/SVG/SvgFactoryTest.java
@@ -4,8 +4,6 @@ import domain.Carport.Carport;
 import domain.Shed.Shed;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 /**
  * CREATED BY mathias @ 15-12-2020 - 22:52
  **/
@@ -13,9 +11,9 @@ class SvgFactoryTest {
 
     @Test
     void drawCarport() {
-        Shed shed = new Shed(250, 360);
-        Carport carport = new Carport(680, 780, Carport.roofType.FLAT, 90);
-        SvgFactory factory = new rect(shed, carport);
+        Shed shed = new Shed(120, 120);
+        Carport carport = new Carport(245, 460, Carport.roofType.FLAT, 90);
+        SvgFactory factory = new svgDraw();
         factory.drawCarport(carport,shed);
     }
 }
